Adapted algorythm for working with unsighned int quantities of initTime and duration
Structure of the class and functions remain the same (in general)

+++ Interface/Develop/note_generator/note_durations.h
@@ -12,7 +12,7 @@
  * Function "writer" is being used for correct displaying.
  */
 struct NoteDur {
-	float noteDur;
+	unsigned int noteDur;
 	std::function<string(string)> func;
 };
